+++ src/app/(routes)/blog/[slug]/components/mdx-components.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 import * as runtime from 'react/jsx-runtime';
 
 type Props = {
