+++ src/components/pageLayout.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 interface PageLayoutProps {
   children: React.ReactNode;
